[BOOKINGSG-9274][GZ] remove deprecated v2 story docs
