--- stories/masonry/masonry.mdx
@@ -1,58 +1,0 @@
-import { Canvas, Meta } from "@storybook/addon-docs/blocks";
-import { Alert } from "src/alert";
-import * as MasonryStories from "./masonry.stories";
-import { PropsTable } from "./props-table";
-
-<Meta of={MasonryStories} />
-
-# Masonry
-
-<Alert type="warning">
-    This component is deprecated and will be removed in a future version.
-    Consider switching to [Layout.ColDiv](/docs/core-layout-coldiv--docs) to
-    manage grid layouts
-</Alert>
-
-## Overview
-
-The component that produces a masonry-tile like layout.
-
-```tsx
-import { Masonry } from "@lifesg/react-design-system/masonry";
-```
-
-The component has 2 sub components:
-
--   `Masonry.Grid` - This holds the wrapper container and it is used to control the number of columns for the `Masonry`
--   `Masonry.Tile` - This element holds the content for each of the grid item
-
-<Canvas of={MasonryStories.Default} />
-
-## Guidelines
-
-While you may specify any number of columns to be present on the `Masonry.Grid`, there are still
-some restrictions to ensure that the design principles are met. These restrictions are:
-
-1. Specifying the number of columns for the large layout (`numOfCols.lg`) will not cascade down to the
-   smaller screen sizes (`md` and `sm`)
-2. However specifying the number of columns for the smaller layouts will be applied to the larger screen sizes
-
-Here are some of the guidelines for the different layouts:
-
-|                 | Large (`lg`) |   Medium (`md`)   |           Small (`sm`) |
-| --------------- | :----------: | :---------------: | ---------------------: |
-| Device widths   |   Desktop    | Tablet (\<1200px) | Mobile phone (\<481px) |
-| Default columns |      -       |         2         |                      1 |
-| Maximum columns |      -       |         -         |                      2 |
-
-<br />
-
-## Layouts for different media widths
-
-Here is an example how you can specify the different layouts for different media widths. (Reduce your viewport size to view)
-
-<Canvas of={MasonryStories.DifferentMediaWidths} />
-
-## Component API
-
-<PropsTable />
