+++ stories/menu/menu.mdx
@@ -15,7 +15,7 @@ displays a menu. It is commonly used in combination with
 
 This module contains:
 
--   **Menu** - a utility component that handles the triggering of the menu, extending [PopoverTrigger](/docs/overlays-popoverv2--docs)
+-   **Menu** - a utility component that handles the triggering of the menu, extending [PopoverTrigger](/docs/overlays-popover--docs)
 -   **Menu.Content** - the main wrapper containing the menu items
 -   **Menu.Section** - groups menu items and optionally shows a section label
 -   **Menu.Item** - a menu item with optional label and sublabel
@@ -33,7 +33,7 @@ import { Menu } from "@lifesg/react-design-system/menu";
 
 ## Resize and flip
 
-As the menu extends [PopoverTrigger](/docs/overlays-popoverv2--docs), resize and
+As the menu extends [PopoverTrigger](/docs/overlays-popover--docs), resize and
 flip behaviour can be configured
 
 <Canvas of={MenuStories.ResizeAndFlip} />
