--- stories/popover-v2/popover.mdx
@@ -1,102 +1,0 @@
-import { Canvas, Meta } from "@storybook/addon-docs/blocks";
-import * as PopoverStories from "./popover.stories";
-import { PropsTable } from "./props-table";
-
-<Meta of={PopoverStories} />
-
-# PopoverV2
-
-## Overview
-
-A component which appears when hovering or clicking on a target element and displays information about a certain element in the page.
-
-This module contains:
-
--   PopoverV2 - the presentational component
--   PopoverTrigger - a utility component that handles triggering of the popover
--   [PopoverInline](/docs/overlays-popoverinline--docs) - a utility component that displays a text or icon-based trigger
-
-```tsx
-import { PopoverTrigger } from "@lifesg/react-design-system/popover-v2";
-```
-
-<Canvas of={PopoverStories.Default} />
-
-<Canvas of={PopoverStories.HoverInteraction} />
-
-### Behaviour
-
-**Appearance**
-
--   The `Popover` component automatically positions itself based on the amount of space available. It will always attempt to display itself at the preferred position by default
--   The `Popover` appears as a modal in mobile viewports
-
-**Keyboard support**
-
-The trigger should be included in the page's tab sequence. This way, users who
-navigate via keyboard can open the popover with `Space`/`Enter`. On open, focus
-will move to the popover. Press `Escape` to dismiss the popover.
-
-The example below illustrates the positioning behaviour when viewed on desktop.
-
-<Canvas of={PopoverStories.Appearance} />
-
-## Default styling
-
-The component comes with basic styling for commonly used markup such as `<strong>` or `<a>`.
-
-<Canvas of={PopoverStories.DefaultStyling} />
-
-## Customisation
-
-You can opt out of the default appearance to display custom content such as menus and dialogs.
-
-<Canvas of={PopoverStories.Customisation} />
-
-## Z-index conflicts
-
-The popover uses z-index to display above elements. It may end up rendering
-below another element with a higher z-index.
-
-The popover is `Modal` and `Overlay`-aware by default; when rendered as a child
-of either component, it will automatically adjust its z-index to remain visible.
-The story below illustrates the default behaviour.
-
-If you still encounter z-index issues, try setting a custom `zIndex` or
-`rootNode`.
-
-<Canvas of={PopoverStories.UsageInOverlay} />
-
-## Resize and Flip
-
-Use `enableResize` to enable resize of the `popoverContent` to fit the remaining
-vertical space of the window and become scrollable.
-
-Use `enableFlip` to disable flipping of the `popoverContent`.
-
-<Canvas of={PopoverStories.ResizeAndFlip} />
-
-With `enableResize`, when `popoverContent` is a render function, additional
-props are provided. You are responsible for applying the `maxHeight` and
-`overflow` values to your custom component to ensure correct sizing and scroll
-behavior.
-
-<Canvas of={PopoverStories.ResizeAndFlipWithCustomContent} />
-
-### Overflow
-
-Use `overflow` with `enableResize` to control the popover content&rsquo;s
-overflow behavior.
-
-<Canvas of={PopoverStories.Overflow} />
-
-## Nested overlays
-
-This popover contains a component with a dropdown. In most cases, the popover
-should be able to coexist with a nested floating element without issue.
-
-<Canvas of={PopoverStories.NestedOverlays} />
-
-## Component API
-
-<PropsTable />
