+++ stories/popover/popover.stories.tsx
@@ -1,42 +1,279 @@
 import type { Meta, StoryObj } from "@storybook/react-webpack5";
 import { useState } from "react";
 import { Button } from "src/button";
-import { Popover } from "src/popover";
+import { Form } from "src/form";
+import { Modal } from "src/modal";
+import { Popover, PopoverTrigger } from "src/popover";
+import { GridDecorator } from "stories/storybook-common";
 
-import { Container, Item, PopoverHOCExample } from "./doc-elements";
+import { CustomPopover, POPOVER_CONTENT } from "./doc-elements";
 
 type Component = typeof Popover;
+type TriggerComponent = typeof PopoverTrigger;
 
 const meta: Meta<Component> = {
-    title: "Deprecated/Popover",
+    title: "Overlays/Popover",
     component: Popover,
 };
 
 export default meta;
 
 export const Default: StoryObj<Component> = {
-    render: () => {
-        return <PopoverHOCExample />;
+    render: (_args) => {
+        return (
+            <PopoverTrigger popoverContent="This is the popover content">
+                <Button.Default>Click me</Button.Default>
+            </PopoverTrigger>
+        );
     },
 };
 
-export const UsingAsAComponent: StoryObj<Component> = {
-    render: () => {
-        const [visible, setVisible] = useState(false);
+export const HoverInteraction: StoryObj<Component> = {
+    render: (_args) => {
         return (
-            <Container>
-                <Item>
-                    <Popover
-                        visible={visible}
-                        onMobileClose={() => setVisible(false)}
-                    >
-                        I am the popover content
-                    </Popover>
-                    <Button.Default onClick={() => setVisible(!visible)}>
-                        Click me to interact
-                    </Button.Default>
-                </Item>
-            </Container>
+            <PopoverTrigger
+                popoverContent="This is the popover content"
+                trigger="hover"
+            >
+                <Button.Default>Hover me</Button.Default>
+            </PopoverTrigger>
+        );
+    },
+};
+
+export const DefaultStyling: StoryObj<Component> = {
+    render: (_args) => {
+        return (
+            <PopoverTrigger
+                popoverContent={
+                    <>
+                        <p>
+                            You can add <strong>bold text</strong> and{" "}
+                            <a>links</a>.
+                        </p>
+                        <br />
+                        <p>You can list bullet points:</p>
+                        <ul>
+                            <li>List item</li>
+                        </ul>
+                        <p>Or display numbered lists:</p>
+                        <ol>
+                            <li>List item</li>
+                        </ol>
+                    </>
+                }
+            >
+                <Button.Default>Click me</Button.Default>
+            </PopoverTrigger>
+        );
+    },
+};
+
+export const Appearance: StoryObj<Component> = {
+    render: (_args) => {
+        return (
+            <>
+                <PopoverTrigger popoverContent="Remains aligned to the left of the screen when screen width is reduced">
+                    <Button.Default>Left</Button.Default>
+                </PopoverTrigger>
+                <PopoverTrigger
+                    popoverContent={
+                        <div style={{ height: "500px" }}>
+                            Remains in view as much as possible. Scroll up and
+                            down to view the repositioning
+                        </div>
+                    }
+                >
+                    <Button.Default>Center</Button.Default>
+                </PopoverTrigger>
+                <PopoverTrigger popoverContent="Remains aligned to the right of the screen when screen width is reduced">
+                    <Button.Default>Right</Button.Default>
+                </PopoverTrigger>
+            </>
+        );
+    },
+    parameters: {
+        layout: "fullscreen",
+    },
+    decorators: [
+        (Story) => (
+            <div
+                style={{
+                    display: "flex",
+                    justifyContent: "space-around",
+                    padding: "3rem 0",
+                }}
+            >
+                <Story />
+            </div>
+        ),
+    ],
+};
+
+export const Customisation: StoryObj<Component> = {
+    render: () => {
+        return (
+            <PopoverTrigger
+                popoverContent={() => (
+                    <div style={{ background: "aliceblue", padding: "1rem" }}>
+                        A custom popover!
+                    </div>
+                )}
+            >
+                <Button.Default>Click me</Button.Default>
+            </PopoverTrigger>
+        );
+    },
+};
+
+export const UsageInOverlay: StoryObj<Component> = {
+    render: () => {
+        const [show, setShow] = useState(false);
+        const [showNested, setShowNested] = useState(false);
+
+        return (
+            <>
+                <Button.Default
+                    onClick={() => {
+                        setShow(true);
+                        setTimeout(() => {
+                            setShowNested(true);
+                        }, 3000);
+                    }}
+                >
+                    Show modal
+                </Button.Default>
+                <Modal show={show} onOverlayClick={() => setShow(false)}>
+                    <PopoverTrigger popoverContent="I'm a tooltip!">
+                        <Button.Default>Toggle tooltip</Button.Default>
+                    </PopoverTrigger>
+                </Modal>
+                <Modal
+                    show={showNested}
+                    onOverlayClick={() => setShowNested(false)}
+                >
+                    <PopoverTrigger popoverContent="I'm a nested tooltip!">
+                        <Button.Default>Toggle nested tooltip</Button.Default>
+                    </PopoverTrigger>
+                </Modal>
+            </>
+        );
+    },
+};
+
+export const ResizeAndFlip: StoryObj<TriggerComponent> = {
+    args: {
+        popoverContent: POPOVER_CONTENT,
+        zIndex: 10,
+    },
+    render: (args) => {
+        return (
+            <>
+                <PopoverTrigger {...args}>
+                    <Button.Default>Click me</Button.Default>
+                </PopoverTrigger>
+                <PopoverTrigger enableResize={true} {...args}>
+                    <Button.Default>Click me</Button.Default>
+                </PopoverTrigger>
+                <PopoverTrigger
+                    enableResize={true}
+                    enableFlip={false}
+                    {...args}
+                >
+                    <Button.Default>Click me</Button.Default>
+                </PopoverTrigger>
+                <PopoverTrigger enableFlip={false} {...args}>
+                    <Button.Default>Click me</Button.Default>
+                </PopoverTrigger>
+            </>
+        );
+    },
+    decorators: [
+        GridDecorator({
+            columns: 4,
+            columnHeaders: [
+                "Resize disabled, Flip enabled (Default)",
+                "Resize and Flip enabled",
+                "Resize enabled, Flip disabled",
+                "Resize and Flip disabled",
+            ],
+        }),
+    ],
+};
+
+export const ResizeAndFlipWithCustomContent: StoryObj<Component> = {
+    render: () => {
+        return (
+            <PopoverTrigger
+                enableResize
+                popoverContent={({ maxHeight, overflow }) => (
+                    <CustomPopover style={{ maxHeight, overflow }} />
+                )}
+            >
+                <Button.Default>Click me</Button.Default>
+            </PopoverTrigger>
+        );
+    },
+};
+
+export const Overflow: StoryObj<TriggerComponent> = {
+    args: {
+        enableResize: true,
+        popoverContent: POPOVER_CONTENT,
+        zIndex: 10,
+    },
+    render: (args) => {
+        return (
+            <>
+                <PopoverTrigger overflow="visible" {...args}>
+                    <Button.Default>Click me</Button.Default>
+                </PopoverTrigger>
+                <PopoverTrigger overflow="hidden" {...args}>
+                    <Button.Default>Click me</Button.Default>
+                </PopoverTrigger>
+                <PopoverTrigger overflow="clip" {...args}>
+                    <Button.Default>Click me</Button.Default>
+                </PopoverTrigger>
+                <PopoverTrigger overflow="scroll" {...args}>
+                    <Button.Default>Click me</Button.Default>
+                </PopoverTrigger>
+                <PopoverTrigger overflow="auto" {...args}>
+                    <Button.Default>Click me</Button.Default>
+                </PopoverTrigger>
+            </>
+        );
+    },
+    decorators: [
+        GridDecorator({
+            columns: 5,
+            columnHeaders: ["Visible", "Hidden", "Clip", "Scroll", "Auto"],
+        }),
+    ],
+};
+
+export const NestedOverlays: StoryObj<Component> = {
+    render: (_args) => {
+        const options = [
+            { label: "Option 1", value: "option1" },
+            { label: "Option 2", value: "option2" },
+            { label: "Option 3", value: "option3" },
+        ];
+
+        return (
+            <PopoverTrigger
+                popoverContent={() => (
+                    <Form.MultiSelect
+                        data-testid="multiselect-filter"
+                        options={options}
+                        listExtractor={(opt) => opt.label}
+                        valueExtractor={(opt) => opt.value}
+                        enableSearch
+                    />
+                )}
+                position="bottom-start"
+            >
+                <Button.Default>Click me</Button.Default>
+            </PopoverTrigger>
         );
     },
 };
