+++ stories/popover/props-table.tsx
@@ -1,136 +1,349 @@
-import {
-    DefaultCol,
-    DescriptionCol,
-    NameCol,
-    Section,
-    Table,
-} from "stories/storybook-common/api-table";
+import { ApiTable } from "stories/storybook-common/api-table";
+import type { ApiTableSectionProps } from "stories/storybook-common/api-table/types";
 import type { TabAttribute } from "stories/storybook-common/tabs";
 import { Tabs } from "stories/storybook-common/tabs";
 
-export const HOCTable = () => (
-    <Table>
-        <Section>PopoverHOCProps</Section>
+export const COMMON_POPOVER_ATTRIBUTES: ApiTableSectionProps["attributes"] = [
+    {
+        name: "popoverContent",
+        description: (
+            <>
+                The content of the <code>Popover</code>. Can be a{" "}
+                <code>string</code>, a <code>JSX.Element</code>, or a function
+                that receives <code>PopoverRenderProps</code>. When using the
+                function form, the returned content can be dynamically styled or
+                sized using the <code>PopoverRenderProps</code> (e.g.,{" "}
+                <code>maxHeight</code>, <code>overflow</code>).
+            </>
+        ),
+        propTypes: [
+            "string",
+            "JSX.Element",
+            "(renderProps: PopoverRenderProps) => React.ReactNode",
+        ],
+        mandatory: true,
+    },
+    {
+        name: "trigger",
+        description: (
+            <>
+                The trigger for the appearance of the <code>Popover</code>
+            </>
+        ),
+        propTypes: [`"click"`, `"hover"`],
+        defaultValue: `"click"`,
+    },
+    {
+        name: "position",
+        description: (
+            <>
+                The visual position of the <code>Popover</code> in relation to
+                its trigger
+            </>
+        ),
+        propTypes: [
+            `"top"`,
+            `"top-start"`,
+            `"top-end"`,
+            `"bottom"`,
+            `"bottom-start"`,
+            `"bottom-end"`,
+            `"left"`,
+            `"left-start"`,
+            `"left-end"`,
+            `"right"`,
+            `"right-start"`,
+            `"right-end"`,
+        ],
+        defaultValue: `"top"`,
+    },
+    {
+        name: "zIndex",
+        description: (
+            <>
+                The custom z-index of the <code>Popover</code>. Try specifying
+                this if you encounter z-index conflicts.
+            </>
+        ),
+        propTypes: ["number"],
+    },
+    {
+        name: "rootNode",
+        description: (
+            <>
+                The root element that hosts the popover element. Try specifying
+                this if <code>zIndex</code> does not work.
+                <br />
+                <br />
+                For example, if the parent of the trigger element has a higher
+                z-index than the popover, the popover may not be visible.
+                Specify the parent here instead so that they share the same
+                stacking context.
+            </>
+        ),
+        propTypes: ["RefObject<HTMLElement>"],
+        defaultValue: (
+            <>
+                document<code>body</code>
+            </>
+        ),
+    },
+    {
+        name: "onPopoverAppear",
+        description: (
+            <>
+                The callback when the <code>Popover</code> appears
+            </>
+        ),
+        propTypes: ["() => void"],
+    },
+    {
+        name: "onPopoverDismiss",
+        description: (
+            <>
+                The callback when the <code>Popover</code> dismisses
+            </>
+        ),
+        propTypes: ["() => void"],
+    },
+    {
+        name: "enableFlip",
+        description: (
+            <>
+                Enables the Popover&rsquo;s flip behavior (vertically) when
+                there is not enough space in the window&rsquo;s viewport for the
+                specified position.
+            </>
+        ),
+        propTypes: ["boolean"],
+        defaultValue: "true",
+    },
+    {
+        name: "overflow",
+        description: (
+            <>
+                Controls how overflow content behaves inside the popover
+                container. Used with <code>enableResize</code>
+            </>
+        ),
+        propTypes: ["visible", "hidden", "clip", "scroll", "auto"],
+        defaultValue: "auto",
+    },
+    {
+        name: "enableResize",
+        description: (
+            <>
+                Enables popover resize to fit the remaining vertical space of
+                the window and contents become scrollable.
+            </>
+        ),
+        propTypes: ["boolean"],
+        defaultValue: "false",
+    },
+    {
+        name: "popoverAriaLabel",
+        description: "The accessible label for the popover content",
+        propTypes: ["string"],
+        defaultValue: "More information",
+    },
+    {
+        name: "triggerOnFocus",
+        description: (
+            <>
+                Opens the popover when the trigger element receives keyboard
+                focus (e.g. via Tab), instead of requiring a click.
+            </>
+        ),
+        propTypes: ["boolean"],
+        defaultValue: "false",
+    },
+    {
+        name: "isModal",
+        description: (
+            <>
+                Controls whether the popover behaves as a modal.
+                <br />
+                <br />
+                When <code>true</code>, focus is trapped within the popover and
+                interaction outside is disabled until it is dismissed.
+                <br />
+                When <code>false</code>, focus may move outside the popover and
+                it behaves as a non-modal element (e.g. navigation menus).
+            </>
+        ),
+        propTypes: ["boolean"],
+        defaultValue: "true",
+    },
+];
 
-        <tr>
-            <NameCol>onPopoverAppear</NameCol>
-            <DescriptionCol propTypes={["() => void"]}>
-                <>
-                    The callback when the <code>Popover</code> appears
-                </>
-            </DescriptionCol>
-            <DefaultCol />
-        </tr>
-        <tr>
-            <NameCol>onPopoverDismiss</NameCol>
-            <DescriptionCol propTypes={["() => void"]}>
-                <>
-                    The callback when the <code>Popover</code> dismisses
-                </>
-            </DescriptionCol>
-            <DefaultCol />
-        </tr>
+const POPOVER_TRIGGER_DATA: ApiTableSectionProps[] = [
+    {
+        attributes: [
+            {
+                name: "className",
+                description: "Class selector for the component",
+                propTypes: ["string"],
+            },
+            {
+                name: "id",
+                description: "The unique identifier for the component",
+                propTypes: ["string"],
+            },
+            {
+                name: "data-testid",
+                description: "The test identifier for the component",
+                propTypes: ["string"],
+            },
+            {
+                name: "children",
+                description: (
+                    <>
+                        The item to interact with to trigger the{" "}
+                        <code>Popover</code>
+                    </>
+                ),
+                propTypes: ["React.ReactNode"],
+                mandatory: true,
+            },
+            {
+                name: "customOffset",
+                description: "How much offset to apply for the popover (in px)",
+                propTypes: ["number"],
+                defaultValue: "16",
+            },
+            {
+                name: "delay",
+                description: (
+                    <>
+                        How much delay for popover to appear/dismiss <br />
+                        <b>Note</b>: When trigger is of type <code>click</code>,
+                        delay is not applied.
+                    </>
+                ),
+                propTypes: ["{ open: number, close: number }"],
+                defaultValue: "{ open: 0, close: 500 }",
+            },
+            ...COMMON_POPOVER_ATTRIBUTES,
+        ],
+    },
+    {
+        name: "PopoverRenderProps",
+        attributes: [
+            {
+                name: "maxHeight",
+                mandatory: false,
+                description:
+                    "Maximum height (in pixels) allowed for the popover content.",
+                propTypes: ["number"],
+            },
+            {
+                name: "overflow",
+                mandatory: false,
+                description: (
+                    <>
+                        Controls how overflow content is handled inside the
+                        popover.
+                    </>
+                ),
+                propTypes: [
+                    '"auto"',
+                    '"scroll"',
+                    '"clip"',
+                    '"hidden"',
+                    '"visible"',
+                ],
+            },
+        ],
+    },
+];
 
-        <Section>PopoverHOCOptionsProps</Section>
-
-        <tr>
-            <NameCol mandatory>content</NameCol>
-            <DescriptionCol propTypes={["string", "JSX.Element"]}>
-                <>
-                    The content of the <code>Popover</code>
-                </>
-            </DescriptionCol>
-            <DefaultCol />
-        </tr>
-        <tr>
-            <NameCol>trigger</NameCol>
-            <DescriptionCol propTypes={[`"click"`, `"hover"`]}>
-                <>
-                    The trigger for the appearance of the <code>Popover</code>
-                </>
-            </DescriptionCol>
-            <DefaultCol>{[`"click"`]}</DefaultCol>
-        </tr>
-        <tr>
-            <NameCol>id</NameCol>
-            <DescriptionCol propTypes={["string"]}>
-                A unique identifier for the component
-            </DescriptionCol>
-            <DefaultCol />
-        </tr>
-        <tr>
-            <NameCol>data-testid</NameCol>
-            <DescriptionCol propTypes={["string"]}>
-                A test identifier for the component
-            </DescriptionCol>
-            <DefaultCol />
-        </tr>
-    </Table>
-);
-
-export const PopoverTable = () => (
-    <Table>
-        <tr>
-            <NameCol mandatory>children</NameCol>
-            <DescriptionCol propTypes={["string", "JSX.Element"]}>
-                <>
-                    The content of the <code>Popover</code>
-                </>
-            </DescriptionCol>
-            <DefaultCol />
-        </tr>
-        <tr>
-            <NameCol>visible</NameCol>
-            <DescriptionCol propTypes={["boolean"]}>
-                <>
-                    The visibility state of the <code>Popover</code>
-                </>
-            </DescriptionCol>
-            <DefaultCol>{["false"]}</DefaultCol>
-        </tr>
-        <tr>
-            <NameCol>id</NameCol>
-            <DescriptionCol propTypes={["string"]}>
-                A unique identifier for the component
-            </DescriptionCol>
-            <DefaultCol />
-        </tr>
-        <tr>
-            <NameCol>className</NameCol>
-            <DescriptionCol propTypes={["string"]}>
-                Class selector for the component
-            </DescriptionCol>
-            <DefaultCol />
-        </tr>
-        <tr>
-            <NameCol>data-testid</NameCol>
-            <DescriptionCol propTypes={["string"]}>
-                A test identifier for the component
-            </DescriptionCol>
-            <DefaultCol />
-        </tr>
-        <tr>
-            <NameCol>onMobileClose</NameCol>
-            <DescriptionCol propTypes={["() => void"]}>
-                <>
-                    The callback when the user dismisses the{" "}
-                    <code>Popover</code>
-                    (modal form) in mobile viewports
-                </>
-            </DescriptionCol>
-            <DefaultCol />
-        </tr>
-    </Table>
-);
+const POPOVER_DATA: ApiTableSectionProps[] = [
+    {
+        attributes: [
+            {
+                name: "className",
+                description: "Class selector for the component",
+                propTypes: ["string"],
+            },
+            {
+                name: "id",
+                description: "The unique identifier for the component",
+                propTypes: ["string"],
+            },
+            {
+                name: "data-testid",
+                description: "The test identifier for the component",
+                propTypes: ["string"],
+            },
+            {
+                name: "children",
+                description: (
+                    <>
+                        The content of the <code>Popover</code>
+                    </>
+                ),
+                propTypes: ["string", "JSX.Element"],
+                mandatory: true,
+            },
+            {
+                name: "visible",
+                description: (
+                    <>
+                        The visibility state of the <code>Popover</code>
+                    </>
+                ),
+                propTypes: ["boolean"],
+            },
+            {
+                name: "onMobileClose",
+                description: (
+                    <>
+                        The callback when the user dismisses the{" "}
+                        <code>Popover</code>
+                        (modal form) in mobile viewports
+                    </>
+                ),
+                propTypes: ["() => void"],
+            },
+            {
+                name: "overflow",
+                description: (
+                    <>
+                        Controls how overflow content behaves inside the popover
+                        container (Desktop only).
+                    </>
+                ),
+                propTypes: ["visible", "hidden", "clip", "scroll", "auto"],
+            },
+            {
+                name: "maxHeight",
+                description: (
+                    <>
+                        Sets the maximum height of the popover container in
+                        pixels. If content exceeds this height, overflow
+                        behavior applies (Desktop only).
+                    </>
+                ),
+                propTypes: ["number"],
+            },
+            {
+                name: "ariaLabel",
+                description: "The accessible label for the popover content",
+                propTypes: ["string"],
+                defaultValue: "More information",
+            },
+        ],
+    },
+];
 
 const PROPS_TABLE_DATA: TabAttribute[] = [
     {
-        title: "withPopover",
-        component: <HOCTable />,
+        title: "PopoverTrigger",
+        component: <ApiTable sections={POPOVER_TRIGGER_DATA} />,
     },
     {
         title: "Popover",
-        component: <PopoverTable />,
+        component: <ApiTable sections={POPOVER_DATA} />,
     },
 ];
