--- stories/tooltip/tooltip.mdx
@@ -1,142 +1,0 @@
-import { Canvas, Meta } from "@storybook/addon-docs/blocks";
-import { Alert } from "src/alert";
-import * as TooltipStories from "./tooltip.stories";
-import { PropsTable } from "./props-table";
-
-<Meta of={TooltipStories} />
-
-# Tooltip
-
-<Alert type="warning">
-    This component is deprecated and will be removed in a future version. Switch
-    to [PopoverV2](/docs/overlays-popoverv2--docs) instead.
-</Alert>
-
-## Overview
-
-A component which appears when hovering or clicking on a target element and
-displays information about a certain element in the page.
-
-```tsx
-import { withTooltip } from "@lifesg/react-design-system/tooltip";
-```
-
-<Canvas of={TooltipStories.Default} />
-
-{/* USAGE */}
-
-## Usage
-
-We recommend using the `withTooltip` higher order component (HOC) to construct the component. Here
-is how you can do it.
-
-### Step 1: Create the trigger component
-
-```tsx
-import { Button, ButtonProps } from "@lifesg/react-design-system/button";
-
-const TriggerComponent = (props: ButtonProps) => (
-    <Button.Default {...props}>Click on me</Button.Default>
-);
-```
-
-### Step 2: Use the HOC function to create a HOC
-
-```tsx
-const WrappedComponent = withTooltip(TriggerComponent, {
-    content: "This is my tooltip",
-});
-```
-
-### Step 3: Compose the component with the HOC
-
-You may choose to handle the visibility of the `Tooltip` within the composed function or
-expose props to allow the usage from parent elements.
-
-```tsx
-export const TooltipHOCExample = () => {
-    const [visible, setVisible] = useState(false);
-
-    return (
-        <div style={{ display: "flex" }}>
-            <WrappedComponent
-                onClick={() => setVisible(!visible)}
-                position="right"
-                tooltipVisible={visible}
-            />
-        </div>
-    );
-};
-```
-
-> Take note that when using a HOC, it is important to
-> **ensure that the declaration of the HOC is not within the functional component**.
-> Doing so would cause certain display animations to not function correctly.
-
-```tsx
-/** Wrong **/
-
-export const WrongExample = () => {
-    const Wrapped = withTooltip(Component, {
-        content: "This is my tooltip",
-    }); // You should not do this
-
-    /** state and event handlers code */
-
-    return (
-        <Container>
-            <Wrapped
-                onClick={handleClick}
-                tooltipVisible={tooltipVisible}
-                position="right"
-            />
-        </Container>
-    );
-};
-
-/** Correct **/
-
-const WrappedCorrectly = withTooltip(Component, {
-    content: "This is my tooltip",
-});
-
-export const CorrectExample = () => {
-    /** state and event handlers code */
-
-    return (
-        <Container>
-            <WrappedCorrectly
-                onClick={handleClick}
-                tooltipVisible={tooltipVisible}
-                position="right"
-            />
-        </Container>
-    );
-};
-```
-
-{/* USING AS A COMPONENT */}
-
-## Using as a component
-
-You also have the freedom to construct the `Tooltip` as a separate component from its trigger.
-
-```tsx
-import { Tooltip } from "@lifesg/react-design-system/tooltip";
-```
-
-<Canvas of={TooltipStories.UsingAsAComponent} />
-
-{/* POSITIONS */}
-
-## Available positions
-
-There are 4 types of positions that the `Tooltip` can appear. You can specify so using the `position` prop.
-
-In this example, click to interact with the buttons.
-
-<Canvas of={TooltipStories.AvailablePositions} />
-
-## Component API
-
-<PropsTable />
