--- stories/v2/introduction.mdx
@@ -1,18 +1,0 @@
-import { Meta } from "@storybook/addon-docs/blocks";
-
-<Meta title="V2/Introduction" />
-
-# V2
-
-This section houses the deprecated modules that are based on the old theme
-system in V2. These modules will be removed in future versions of the design
-system.
-
-The new design tokens are more powerful and cater for a wider variety of use
-cases. You are encouraged to upgrade as soon as possible.
-
-## Migrating to V3
-
-Refer to the
-[wiki](https://github.com/LifeSG/react-design-system/wiki/Migrating-to-V3) for
-more detailed migration steps.
