--- stories/v2/media-query.mdx
@@ -1,95 +1,0 @@
-import { Meta } from "@storybook/addon-docs/blocks";
-import { DocAlert } from "stories/storybook-common";
-
-<Meta title="V2/Media Query" />
-
-# Media Query
-
-<DocAlert>
-    This module was deprecated and replaced by
-    [Breakpoint](/docs/foundations-breakpoint-introduction--docs) tokens in
-    v3.0.0.
-</DocAlert>
-
-## Overview
-
-To target different media widths, you can make use of this component.
-
-```tsx
-import { V2_MediaQuery } from "@lifesg/react-design-system/v2_media";
-```
-
-Here are the list of media width breakpoints that this module supports.
-
-| Name        |   Size Ranges   | Breakpoint value |
-| ----------- | :-------------: | :--------------: |
-| `mobileS`   |    \<= 320px    |      320px       |
-| `mobileM`   |  321px - 375px  |      375px       |
-| `mobileL`   |  376px - 480px  |      480px       |
-| `tablet`    | 481px - 1199px  |      1199px      |
-| `desktopM`  | 1200px - 1380px |      1399px      |
-| `desktopL`  | 1400px - 1999px |      1999px      |
-| `desktop4k` |    >= 2000px    |      3840px      |
-
-<br />
-
-## Usage
-
-You can either target styles using the `max-width` or `min-width` breakpoints. Depending on your approach towards a desktop-first or mobile-first implementation.
-
-## Max Width
-
-```tsx
-import styled from "styled-components";
-import { V2_MediaQuery } from "@lifesg/react-design-system/v2_media";
-
-const Components = styled.div`
-	${V2_MediaQuery.MaxWidth.mobileL} {
-		// Your style
-	}
-`;
-
-// Resolves to
-@media screen and (max-width: 480px) {
-
-}
-```
-
-## Min Width
-
-```tsx
-import styled from "styled-components";
-import { V2_MediaQuery } from "@lifesg/react-design-system/v2_media";
-
-const Components = styled.div`
-	${V2_MediaQuery.MinWidth.mobileL} {
-		// Your style
-	}
-`;
-
-// Resolves to
-@media screen and (min-width: 480px) {
-
-}
-```
-
-## Using media width values
-
-```tsx
-import { V2_MediaWidths } from "@lifesg/react-design-system/v2_media";
-```
-
-You can import `V2_MediaWidths` and use the value as required. It will return the various breakpoint values.
-
-```tsx
-import { V2_MediaWidths } from "@lifesg/react-design-system/v2_media";
-
-export const MyComponent = () => {
-    const { innerWidth } = window;
-    if (innerWidth > V2_MediaWidths.mobileL) {
-        // render something
-    } else {
-        // render something else
-    }
-};
-```
