--- stories/v2_color/a-color.mdx
@@ -1,62 +1,0 @@
-import { Meta } from "@storybook/addon-docs/blocks";
-import { DocAlert } from "stories/storybook-common";
-
-<Meta title="V2/Color/Introduction" />
-
-# Color
-
-<DocAlert>
-    This module was deprecated and replaced by
-    [Colour](/docs/foundations-colours-introduction--docs) tokens in v3.0.0.
-</DocAlert>
-
-## Overview
-
-Colors provide a visual identity to the user interface in the LifeSG and BookingSG apps.
-
-```tsx
-import { V2_Color } from "@lifesg/react-design-system/v2_color";
-```
-
-## Usage
-
-You can use the `Color` component to declare colours within style declarations.
-
-```tsx
-import { V2_Color } from "@lifesg/react-design-system/v2_Color";
-import styled, { css } from "styled-components";
-
-const Container = styled.div`
-    background: ${V2_Color.Primary};
-
-    // If you extract props, remember to wrap nested interpolations with the css helper
-    ${(props) => {
-        return css`
-            color: ${V2_Color.Secondary};
-        `;
-    }}
-
-    // Otherwise, you have to pass the props down manually (not recommended)
-	${(props) => {
-        return `
-			color: ${V2_Color.Secondary(props)};
-		`;
-    }}
-`;
-```
-
-If you require the color as a constant value, you could do this
-
-```tsx
-import { V2_Color } from "@lifesg/react-design-system/v2_color";
-import { BaseTheme } from "@lifesg/react-design-system/theme";
-
-const PrimaryColor = V2_Color.Primary({ theme: BaseTheme });
-```
-
-## Collections
-
-[Base Set](/docs/v2-color-lifesg-color--docs) |
-[BookingSG Set](/docs/v2-color-bookingsg-color--docs) |
-[MyLegacy Set](/docs/v2-color-mylegacy-color--docs) |
-[RBS Set](/docs/v2-color-rbs-color--docs)
