--- stories/v2_color/base-color.mdx
@@ -1,51 +1,0 @@
-import { Meta, Unstyled } from "@storybook/addon-docs/blocks";
-import { BaseColorSet } from "src/v2_spec/color-spec/base-color-set";
-import { V2_Text } from "src/v2_text";
-import { PreviewBox } from "../storybook-common";
-import { ColorPalette } from "./doc-color-palette";
-
-<Meta title="V2/Color/LifeSG Color" />
-
-# LifeSG Colors
-
-## Overview
-
-This will be the palette used when the `colorScheme` is `"base"`.
-
-```tsx
-const theme: ThemeSpec = {
-    colorScheme: "base",
-    // ...other specifications
-};
-```
-
-## Colors
-
-<Unstyled>
-    <ColorPalette set={BaseColorSet.Brand} category="Brand" />
-    <ColorPalette
-        set={{
-            Primary: BaseColorSet.Primary,
-            PrimaryDark: BaseColorSet.PrimaryDark,
-            Secondary: BaseColorSet.Secondary,
-        }}
-    />
-    <ColorPalette set={BaseColorSet.Neutral} category="Neutral" />
-    <ColorPalette set={BaseColorSet.Accent.Dark} category="Accent.Dark" />
-    <ColorPalette set={BaseColorSet.Accent.Light} category="Accent.Light" />
-</Unstyled>
-
-<br />
-
-<V2_Text.H3>Validation</V2_Text.H3>
-
-<PreviewBox>
-  `Color.Validation.<color>`
-</PreviewBox>
-
-<Unstyled>
-    <ColorPalette set={BaseColorSet.Validation.Green} category="Green" />
-    <ColorPalette set={BaseColorSet.Validation.Orange} category="Orange" />
-    <ColorPalette set={BaseColorSet.Validation.Red} category="Red" />
-    <ColorPalette set={BaseColorSet.Validation.Blue} category="Blue" />
-</Unstyled>
