--- stories/v2_color/bsg-color.mdx
@@ -1,54 +1,0 @@
-import { Meta, Unstyled } from "@storybook/addon-docs/blocks";
-import { BookingSGColorSet } from "src/v2_spec/color-spec/bookingsg-color-set";
-import { V2_Text } from "src/v2_text";
-import { PreviewBox } from "../storybook-common";
-import { ColorPalette } from "./doc-color-palette";
-
-<Meta title="V2/Color/BookingSG Color" />
-
-# BookingSG Colors
-
-## Overview
-
-This will be the palette used when the `colorScheme` is `"bookingsg"`.
-
-```tsx
-const theme: ThemeSpec = {
-    colorScheme: "bookingsg",
-    // ...other specifications
-};
-```
-
-## Colors
-
-<Unstyled>
-    <ColorPalette set={BookingSGColorSet.Brand} category="Brand" />
-    <ColorPalette
-        set={{
-            Primary: BookingSGColorSet.Primary,
-            PrimaryDark: BookingSGColorSet.PrimaryDark,
-            Secondary: BookingSGColorSet.Secondary,
-        }}
-    />
-    <ColorPalette set={BookingSGColorSet.Neutral} category="Neutral" />
-    <ColorPalette set={BookingSGColorSet.Accent.Dark} category="Accent.Dark" />
-    <ColorPalette
-        set={BookingSGColorSet.Accent.Light}
-        category="Accent.Light"
-    />
-</Unstyled>
-
-<br />
-
-<V2_Text.H3>Validation</V2_Text.H3>
-
-<PreviewBox>
-  `Color.Validation.<color>`
-</PreviewBox>
-
-<Unstyled>
-    <ColorPalette set={BookingSGColorSet.Validation.Green} category="Green" />
-    <ColorPalette set={BookingSGColorSet.Validation.Orange} category="Orange" />
-    <ColorPalette set={BookingSGColorSet.Validation.Red} category="Red" />
-    <ColorPalette set={BookingSGColorSet.Validation.Blue} category="Blue" />
-</Unstyled>
