--- stories/v2_color/ccube-color.mdx
@@ -1,51 +1,0 @@
-import { Meta, Unstyled } from "@storybook/addon-docs/blocks";
-import { CCubeColorSet } from "src/v2_spec/color-spec/ccube-color-set";
-import { V2_Text } from "src/v2_text";
-import { PreviewBox } from "../storybook-common";
-import { ColorPalette } from "./doc-color-palette";
-
-<Meta title="V2/Color/CCube Color" />
-
-# CCube Colors
-
-## Overview
-
-This will be the palette used when the `colorScheme` is `"ccube"`.
-
-```tsx
-const theme: ThemeSpec = {
-    colorScheme: "ccube",
-    // ...other specifications
-};
-```
-
-## Colors
-
-<Unstyled>
-    <ColorPalette set={CCubeColorSet.Brand} category="Brand" />
-    <ColorPalette
-        set={{
-            Primary: CCubeColorSet.Primary,
-            PrimaryDark: CCubeColorSet.PrimaryDark,
-            Secondary: CCubeColorSet.Secondary,
-        }}
-    />
-    <ColorPalette set={CCubeColorSet.Neutral} category="Neutral" />
-    <ColorPalette set={CCubeColorSet.Accent.Dark} category="Accent.Dark" />
-    <ColorPalette set={CCubeColorSet.Accent.Light} category="Accent.Light" />
-</Unstyled>
-
-<br />
-
-<V2_Text.H3>Validation</V2_Text.H3>
-
-<PreviewBox>
-  `Color.Validation.<color>`
-</PreviewBox>
-
-<Unstyled>
-    <ColorPalette set={CCubeColorSet.Validation.Green} category="Green" />
-    <ColorPalette set={CCubeColorSet.Validation.Orange} category="Orange" />
-    <ColorPalette set={CCubeColorSet.Validation.Red} category="Red" />
-    <ColorPalette set={CCubeColorSet.Validation.Blue} category="Blue" />
-</Unstyled>
