--- stories/v2_color/mylegacy-color.mdx
@@ -1,51 +1,0 @@
-import { Meta, Unstyled } from "@storybook/addon-docs/blocks";
-import { MyLegacyColorSet } from "src/v2_spec/color-spec/mylegacy-color-set";
-import { V2_Text } from "src/v2_text";
-import { PreviewBox } from "../storybook-common";
-import { ColorPalette } from "./doc-color-palette";
-
-<Meta title="V2/Color/MyLegacy Color" />
-
-# MyLegacy Colors
-
-## Overview
-
-This will be the palette used when the `colorScheme` is `"mylegacy"`.
-
-```tsx
-const theme: ThemeSpec = {
-    colorScheme: "mylegacy",
-    // ...other specifications
-};
-```
-
-## Colors
-
-<Unstyled>
-    <ColorPalette set={MyLegacyColorSet.Brand} category="Brand" />
-    <ColorPalette
-        set={{
-            Primary: MyLegacyColorSet.Primary,
-            PrimaryDark: MyLegacyColorSet.PrimaryDark,
-            Secondary: MyLegacyColorSet.Secondary,
-        }}
-    />
-    <ColorPalette set={MyLegacyColorSet.Neutral} category="Neutral" />
-    <ColorPalette set={MyLegacyColorSet.Accent.Dark} category="Accent.Dark" />
-    <ColorPalette set={MyLegacyColorSet.Accent.Light} category="Accent.Light" />
-</Unstyled>
-
-<br />
-
-<V2_Text.H3>Validation</V2_Text.H3>
-
-<PreviewBox>
-  `Color.Validation.<color>`
-</PreviewBox>
-
-<Unstyled>
-    <ColorPalette set={MyLegacyColorSet.Validation.Green} category="Green" />
-    <ColorPalette set={MyLegacyColorSet.Validation.Orange} category="Orange" />
-    <ColorPalette set={MyLegacyColorSet.Validation.Red} category="Red" />
-    <ColorPalette set={MyLegacyColorSet.Validation.Blue} category="Blue" />
-</Unstyled>
