--- stories/v2_color/oneservice-color.mdx
@@ -1,58 +1,0 @@
-import { Meta, Unstyled } from "@storybook/addon-docs/blocks";
-import { OneServiceColorSet } from "src/v2_spec/color-spec/oneservice-color-set";
-import { V2_Text } from "src/v2_text";
-import { PreviewBox } from "../storybook-common";
-import { ColorPalette } from "./doc-color-palette";
-
-<Meta title="V2/Color/OneService Color" />
-
-# OneService Colors
-
-## Overview
-
-This will be the palette used when the `colorScheme` is `"oneservice"`.
-
-```tsx
-const theme: ThemeSpec = {
-    colorScheme: "oneservice",
-    // ...other specifications
-};
-```
-
-## Colors
-
-<Unstyled>
-  <ColorPalette set={OneServiceColorSet.Brand} category="Brand" />
-
-<ColorPalette
-    set={{
-        Primary: OneServiceColorSet.Primary,
-        PrimaryDark: OneServiceColorSet.PrimaryDark,
-        Secondary: OneServiceColorSet.Secondary,
-    }}
-/>
-
-<ColorPalette set={OneServiceColorSet.Neutral} category="Neutral" />
-
-<ColorPalette set={OneServiceColorSet.Accent.Dark} category="Accent.Dark" />
-
-  <ColorPalette set={OneServiceColorSet.Accent.Light} category="Accent.Light" />
-</Unstyled>
-
-<br />
-
-<V2_Text.H3>Validation</V2_Text.H3>
-
-<PreviewBox>
-  `Color.Validation.<color>`
-</PreviewBox>
-
-<Unstyled>
-  <ColorPalette set={OneServiceColorSet.Validation.Green} category="Green" />
-
-<ColorPalette set={OneServiceColorSet.Validation.Orange} category="Orange" />
-
-<ColorPalette set={OneServiceColorSet.Validation.Red} category="Red" />
-
-  <ColorPalette set={OneServiceColorSet.Validation.Blue} category="Blue" />
-</Unstyled>
