--- stories/v2_color/rbs-color.mdx
@@ -1,58 +1,0 @@
-import { Meta, Unstyled } from "@storybook/addon-docs/blocks";
-import { RBSColorSet } from "src/v2_spec/color-spec/rbs-color-set";
-import { V2_Text } from "src/v2_text";
-import { PreviewBox } from "../storybook-common";
-import { ColorPalette } from "./doc-color-palette";
-
-<Meta title="V2/Color/RBS Color" />
-
-# RBS Colors
-
-## Overview
-
-This will be the palette used when the `colorScheme` is `"rbs"`.
-
-```tsx
-const theme: ThemeSpec = {
-    colorScheme: "rbs",
-    // ...other specifications
-};
-```
-
-## Colors
-
-<Unstyled>
-  <ColorPalette set={RBSColorSet.Brand} category="Brand" />
-
-<ColorPalette
-    set={{
-        Primary: RBSColorSet.Primary,
-        PrimaryDark: RBSColorSet.PrimaryDark,
-        Secondary: RBSColorSet.Secondary,
-    }}
-/>
-
-<ColorPalette set={RBSColorSet.Neutral} category="Neutral" />
-
-<ColorPalette set={RBSColorSet.Accent.Dark} category="Accent.Dark" />
-
-  <ColorPalette set={RBSColorSet.Accent.Light} category="Accent.Light" />
-</Unstyled>
-
-<br />
-
-<V2_Text.H3>Validation</V2_Text.H3>
-
-<PreviewBox>
-  `Color.Validation.<color>`
-</PreviewBox>
-
-<Unstyled>
-  <ColorPalette set={RBSColorSet.Validation.Green} category="Green" />
-
-<ColorPalette set={RBSColorSet.Validation.Orange} category="Orange" />
-
-<ColorPalette set={RBSColorSet.Validation.Red} category="Red" />
-
-  <ColorPalette set={RBSColorSet.Validation.Blue} category="Blue" />
-</Unstyled>
