--- stories/v2_layout/col-div.mdx
@@ -1,64 +1,0 @@
-import { Canvas, Meta } from "@storybook/addon-docs/blocks";
-import * as ColDivStories from "./col-div.stories";
-import { PropsTable } from "./col-div-props-table";
-
-<Meta of={ColDivStories} />
-
-# Column Divs
-
-## Overview
-
-A div that conforms to the grid layout of `Layout.Container`.
-
-```tsx
-import { V2_Layout } from "@lifesg/react-design-system/v2_layout";
-```
-
-## Anatomy
-
-The Column Divs allows specifying the column widths for each viewport pertaining
-to the specifications of the [grid system](/docs/v2-layout--docs#grid-layout).
-**By default, the component will only take up 1 column in the grid regardless of
-the viewport.**
-
-There are 3 levels of viewport specific props that you can set. Here is a rule
-table to denote the priority of how the viewport specific props are applied.
-
-| Viewport | `mobileCols` | `tabletCols` | `desktopCols` | To use        |
-| -------- | ------------ | ------------ | ------------- | ------------- |
-| Desktop  | Defined      | -            | -             | `mobileCols`  |
-|          | Defined      | Defined      | -             | `tabletCols`  |
-|          | Defined      | Defined      | Defined       | `desktopCols` |
-|          | Defined      | -            | Defined       | `desktopCols` |
-|          | -            | -            | Defined       | `desktopCols` |
-| Tablet   | Defined      | -            | -             | `mobileCols`  |
-|          | Defined      | Defined      | -             | `tabletCols`  |
-|          | Defined      | Defined      | Defined       | `tabletCols`  |
-|          | Defined      | -            | Defined       | `mobileCols`  |
-|          | -            | -            | Defined       | Default 1 col |
-| Mobile   | Defined      | -            | -             | `mobileCols`  |
-|          | Defined      | Defined      | -             | `mobileCols`  |
-|          | Defined      | Defined      | Defined       | `mobileCols`  |
-|          | Defined      | -            | Defined       | `mobileCols`  |
-|          | -            | -            | Defined       | Default 1 col |
-|          | -            | Defined      | -             | Default 1 col |
-
-Here is an example of how we can specify the different column widths that are mobile
-responsive.
-
-> Enter the Canvas mode to play around with the different viewports
-
-<Canvas of={ColDivStories.ColumnDivs} />
-
-## Specifying ranges
-
-Instead of a fixed number of columns, you can also specify the start and end
-columns for the div for each viewports.
-
-> Enter the Canvas mode to play around with the different viewports
-
-<Canvas of={ColDivStories.SpecifyingRanges} />
-
-## Component API
-
-<PropsTable />
