--- stories/v2_layout/layout.mdx
@@ -1,199 +1,0 @@
-import { Canvas, Meta } from "@storybook/addon-docs/blocks";
-import { DocAlert } from "stories/storybook-common";
-import { V2_Layout } from "src/v2_layout";
-import { V2_Text } from "src/v2_text";
-import { Label, PaddingLabel } from "./doc-elements";
-import * as LayoutStories from "./layout.stories";
-import { PropsTable } from "./props-table";
-
-<Meta of={LayoutStories} />
-
-# Layout
-
-<DocAlert>
-    This module was deprecated and replaced by [Layout](/docs/core-layout--docs)
-    components in v3.0.0.
-</DocAlert>
-
-## Overview
-
-The layout components handle responsive displays for its child elements.
-
-```tsx
-import { V2_Layout } from "@lifesg/react-design-system/v2_layout";
-```
-
-## Anatomy
-
-A standard layout comprises of 2 components:
-
--   `Section` the component that represents the `<section>` tag
--   `Container` the component that represents the child `<div>` tag
-
-### Section
-
-This is the wrapper component that uses the `<section>` under the hood.
-It's style specifications are as such:
-
-| Attribute          | Value    |
-| ------------------ | -------- |
-| width              | `100%`   |
-| horizontal padding | `1.5rem` |
-| vertical padding   | `0`      |
-
-### Container
-
-This is the child component that uses a `<div>` under the hood.
-This component **comes with a horizontal padding** on certain viewports and
-enforces a maximum width based on the viewport. The maximum width values are
-as follow:
-
-> **Note**: The horizontal padding is meant to be aligned with the
-> [Singapore Government Design System](https://designsystem.tech.gov.sg/)
-
-| Viewport                        | Max width          | Horizontal Padding |
-| ------------------------------- | ------------------ | ------------------ |
-| `desktop4k`, `desktopL`         | `1320px`           | `0.75rem`          |
-| `desktopM`                      | `1140px`           | `0.75rem`          |
-| `tablet`                        | `720px`            | `0.75rem`          |
-| `mobileL`, `mobileM`, `mobileS` | Takes `100%` width | `0`                |
-
-<br />
-
-{/* DIAGRAM */}
-
-<div style={{ padding: "3rem 0" }}>
-    <V2_Layout.Section
-        style={{
-            padding: "2.5rem 1.5rem",
-            background: "#F9B5B2",
-            display: "flex",
-            justifyContent: "center",
-        }}
-    >
-        <Label>Section</Label>
-        <PaddingLabel>1.5rem</PaddingLabel>
-        <PaddingLabel direction="right">1.5rem</PaddingLabel>
-        <V2_Layout.Container
-            style={{
-                background: "#FDDDD7",
-                display: "flex",
-                justifyContent: "center",
-            }}
-        >
-            <Label>Container</Label>
-            <div
-                style={{
-                    width: "100%",
-                    height: "50px",
-                }}
-            />
-        </V2_Layout.Container>
-    </V2_Layout.Section>
-    <br />
-    <V2_Text.XSmall>
-        <em>
-            Note: <strong>Section</strong> does not come with vertical padding.
-            This is for illustration purposes.
-        </em>
-    </V2_Text.XSmall>
-</div>
-
-{/* STORY */}
-
-## Usage
-
-## General usage
-
-To allow ease of use and not having to compose the two components, we have provided the `Content` that comes with the `Section` and `Container`. This is the **recommended usage** and should fit most use cases.
-
-<Canvas of={LayoutStories.GeneralUsage} />
-
-## Customising the style
-
-There are two ways to customize the style of the `Layout` elements.
-
-### Nesting styles
-
-If you are using `Layout.Content`, you are able to modify the styles by extending the component.
-
-```tsx
-import { V2_Layout } from "@lifesg/react-design-system/v2_layout";
-import styled from "styled-components";
-
-// Method 1: Using the data-id property
-export const ModifiedContent = styled(Layout.Content)`
-    /* Modifying styles here will modify the Section component */
-
-    [data-id="container"] {
-        /* Modifying styles here will modify the Container component */
-    }
-`;
-
-// Method 2: Using the component itself
-export const ModifiedContent2 = styled(V2_Layout.Content)`
-    /* Modifying styles here will modify the Section component */
-
-    ${V2_Layout.Container} {
-        /* Modifying styles here will modify the Container component */
-    }
-`;
-```
-
-### Breaking up V2_Layout.Content
-
-You can break up the `Section` and `Container` separately and style them separately for more freedom. This is useful
-if you intend to add complex styles or other elements inside `Section`.
-
-```tsx
-import { V2_Layout } from "@lifesg/react-design-system/layout";
-import styled from "styled-components";
-
-const ModifiedSection = styled(V2_Layout.Section)`
-    /* Your styles here */
-`;
-
-const ModifiedContainer = styled(V2_Layout.Container)`
-    /* Your styles here */
-`;
-
-// Composition of the components
-export const MyComponent = () => (
-    <ModifiedSection>
-        <ModifiedContainer>
-            <div>The content here</div>
-        </ModifiedContainer>
-    </ModifiedSection>
-);
-```
-
-## Split usage
-
-As mentioned above, you can separately compose the `Section` and `Container` components.
-
-<Canvas of={LayoutStories.SplitUsage} />
-
-## Flex column layout
-
-You can arrange content vertically with the `type` prop set to `flex-column`.
-
-<Canvas of={LayoutStories.FlexColumnLayout} />
-
-## Grid layout
-
-You can make use of the `type` prop to specify if the layout will use the [CSS Grid](https://css-tricks.com/snippets/css/complete-guide-grid/).
-
-<Canvas of={LayoutStories.GridLayout} />
-
-## About the Grid
-
-The design system uses a **12 column** grid layout, with a column gutter of `2rem` for desktop viewports. The other viewport sizes are as such:
-
--   Tablet **8 columns**, `1.5rem` gutter
--   Mobile **4 columns**, `1rem` gutter
-
-<br />
-
-## Component API
-
-<PropsTable />
