--- stories/v2_text-list/text-list.mdx
@@ -1,78 +1,0 @@
-import { Canvas, Meta } from "@storybook/addon-docs/blocks";
-import * as TextListStories from "./text-list.stories";
-import { PropsTable } from "./props-table";
-
-<Meta of={TextListStories} />
-
-# TextList
-
-## Overview
-
-Renders a group of related contents within a list format.
-
-```tsx
-import { V2_TextList } from "@lifesg/react-design-system/v2-text-list";
-```
-
-## Unordered List
-
-This list is used when order does not matter and it does not follow a numbering system for each list item.
-
-You can also choose the bullet type from the selection. Types include `disc`, `circle`, `square`, `none`.
-
-<Canvas of={TextListStories.Unordered} />
-
-## Ordered List
-
-This list is used when order matters and it follows a numbering system for each list item.
-
-You can also customize counter type and the counter separator as well. Counter types
-include `lower-alpha`, `decimal`, `lower-roman`.
-
-<Canvas of={TextListStories.Ordered} />
-
-## Nested Lists
-
-<Canvas of={TextListStories.Nested} />
-
-## Other features
-
-Here are other examples the list can do.
-
-<Canvas of={TextListStories.OtherFeatures} />
-
-## Advanced Usage
-
-Under some circumstances, you might want to customise the counter to suit your use case.
-
-<Canvas of={TextListStories.AdvancedUsage} />
-
-You can do so by overriding the default styling.
-
-```tsx
-export const CustomOrderedList = styled(V2_TextList.Ol)`
-    margin-left: 2rem;
-
-    li {
-        margin-left: 1rem;
-    }
-
-    li:before {
-        content: counters(list, ".") ".";
-        left: -3rem;
-    }
-
-    ${V2_MediaQuery.MaxWidth.mobileL} {
-        li {
-            margin-left: 1rem;
-        }
-        li:before {
-            left: -3rem;
-        }
-    }
-`;
-```
-
-## Component API
-
-<PropsTable />
