--- stories/v2_text/a-intro.mdx
@@ -1,44 +1,0 @@
-import { Canvas, Meta, Story } from "@storybook/addon-docs/blocks";
-import { DocAlert } from "stories/storybook-common";
-import * as TextStories from "./a-intro.stories";
-import { PropsTable } from "./props-table";
-
-<Meta title="V2/Text/Introduction" />
-
-# Text
-
-<DocAlert>
-    This module was deprecated and replaced by
-    [Typography](/docs/core-typography--docs) components in v3.0.0.
-</DocAlert>
-
-## Overview
-
-The component that is used for headings, body text, links and more.
-
-```tsx
-import { V2_Text } from "@lifesg/react-design-system/v2_text";
-```
-
-The `Text` component comes with 2 categories of components:
-
--   [Static text (Display, Heading, Body)](#static-text)
--   [Hyperlinks](#hyperlinks)
-
-<br />
-
-## Static Text
-
-<Story of={TextStories.Styles} />
-
-## Hyperlinks
-
-<Canvas of={TextStories.Hyperlinks} />
-
-As we recommend providing an indicator for links leading to external sites, you are required to use the `external` prop to allow for the indicator to appear.
-
-<Canvas of={TextStories.ExternalHyperlinks} />
-
-## Component API
-
-<PropsTable />
