--- stories/v2_text/b-advanced.mdx
@@ -1,77 +1,0 @@
-import { Canvas, Meta, Unstyled } from "@storybook/addon-docs/blocks";
-import { V2_Text } from "src/v2_text";
-import { PreviewBox } from "../storybook-common";
-import * as TextStories from "./b-advanced.stories";
-
-<Meta title="V2/Text/Advanced Usage" />
-
-# Advanced Usage
-
-## Using a combination of styles
-
-In certain use cases, you might need to use multiple `Text` components to achieve the desired result.
-
-### Example 1
-
-Including a hyperlink within a set of text.
-
-<Canvas of={TextStories.Combined} />
-
-### Example 2
-
-Including different weights within a set of text. All you need to do is to specify the `inline` prop
-for the child components.
-
-<Canvas of={TextStories.Inline} />
-
-## Paragraph specification
-
-You can include a paragraph styled spacing between text blocks by specifying the `paragraph` prop.
-
-<Canvas of={TextStories.Paragraph} />
-
-Comparing to if you did not specify it.
-
-<Unstyled>
-  <PreviewBox>
-    <V2_Text.Body>
-      Lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi
-      euismod quam eget ex tincidunt dapibus. Donec vitae leo vehicula,
-      fermentum urna vitae, gravida ex.
-    </V2_Text.Body>
-
-    <V2_Text.Body>
-      Aenean imperdiet faucibus velit, eu maximus libero facilisis ut.
-      Donec nulla nisi, fermentum eget lorem at, feugiat ultricies ex.
-      Aliquam volutpat nibh non suscipit rhoncus.
-    </V2_Text.Body>
-
-  </PreviewBox>
-</Unstyled>
-
-## Switching underlying component
-
-In some cases, you might require to change the underlying component to be
-rendered. You may do so using the `as` prop.
-
-> E.g. `Text.Body` renders a `<p>` but if you can specify another element to be
-> rendered instead. Inspect the element to find out
-
-<Canvas of={TextStories.RenderAs} />
-
-## CSS styling
-
-Use `TextStyleHelper.getTextStyle()` to apply the relevant font properties to
-an element for the specified size and weight. This is useful for containers that
-contain text children but do not require a text element.
-
-```ts
-import { TextStyleHelper } from "@lifesg/react-design-system/text";
-
-export const ButtonContainer = styled.button`
-    ${TextStyleHelper.getTextStyle("Body", "semibold")}
-    padding: 0.5rem;
-`;
-```
-
-<Canvas of={TextStories.TextStyleHelper} />
