--- stories/v2_text/c-base-collection.mdx
@@ -1,39 +1,0 @@
-import { Meta, Unstyled } from "@storybook/addon-docs/blocks";
-import { V2_Text } from "src/v2_text";
-import { BaseTextStyleSet } from "src/v2_spec/text-spec/base-text-style-set";
-import { FontDisplay, FontSizeDisplay } from "./collection-doc-elements";
-import { V2_BaseTheme } from "../../src";
-import { ThemeProvider } from "styled-components";
-
-<Meta title="V2/Text/Base Style" />
-
-# Base Style
-
-## Overview
-
-This will be the collection used when the `textStyleScheme` is `base`.
-
-```tsx
-const theme: ThemeSpec = {
-    textStyleScheme: "base",
-    // ...other specifications
-};
-```
-
-## Font Family
-
-<V2_Text.D1>Open Sans</V2_Text.D1>
-
-<Unstyled>
-    <ThemeProvider theme={V2_BaseTheme}>
-        <FontDisplay />
-    </ThemeProvider>
-</Unstyled>
-
-## Text Styles
-
-<Unstyled>
-    <ThemeProvider theme={V2_BaseTheme}>
-        <FontSizeDisplay textStyles={BaseTextStyleSet} />
-    </ThemeProvider>
-</Unstyled>
