--- stories/v2_text/d-oneservice-collection.mdx
@@ -1,42 +1,0 @@
-import { Meta, Unstyled } from "@storybook/addon-docs/blocks";
-import { V2_Text } from "src/v2_text";
-import { OneServiceTextStyleSet } from "src/v2_spec/text-spec/oneservice-text-style-set";
-import { FontDisplay, FontSizeDisplay } from "./collection-doc-elements";
-import { V2_OneServiceTheme } from "../../src";
-import { ThemeProvider } from "styled-components";
-
-<Meta title="V2/Text/OneService Style" />
-
-# OneService Style
-
-## Overview
-
-This will be the collection used when the `textStyleScheme` is `oneservice`.
-
-```tsx
-const theme: ThemeSpec = {
-    textStyleScheme: "oneservice",
-    // ...other specifications
-};
-```
-
-## Font Family
-
-<V2_Text.D1>Open Sans</V2_Text.D1>
-
-<Unstyled>
-    <ThemeProvider theme={V2_OneServiceTheme}>
-        <FontDisplay />
-    </ThemeProvider>
-</Unstyled>
-
-## Text Styles
-
-<Unstyled>
-    <ThemeProvider theme={V2_OneServiceTheme}>
-        <FontSizeDisplay
-            textStyles={OneServiceTextStyleSet}
-            theme={V2_OneServiceTheme}
-        />
-    </ThemeProvider>
-</Unstyled>
