--- stories/v2_text/e-plus-jakarta-sans-collection.mdx
@@ -1,45 +1,0 @@
-import { Meta, Unstyled } from "@storybook/addon-docs/blocks";
-import { V2_Text } from "src/v2_text";
-import { BasePlusSansJakartaStyleSet } from "src/v2_spec/text-spec/base-plus-jakarta-sans-text-style-set";
-import { FontDisplay, FontSizeDisplay } from "./collection-doc-elements";
-import { V2_BaseTheme } from "../../src";
-import { ThemeProvider } from "styled-components";
-
-<Meta title="V2/Text/PlusSansJakarta Style" />
-
-# Plus Jakarta Sans Style
-
-## Overview
-
-This will be the collection used when the `textStyleScheme` is `plusJakartaSans`.
-
-Note that the _font ligatures_ is **_disabled_** by default for this `textStyleScheme`.
-
-```tsx
-const theme: ThemeSpec = {
-    textStyleScheme: "plusJakartaSans",
-    // ...other specifications
-};
-```
-
-## Font Family
-
-<V2_Text.D1>Plus Jakarta Sans</V2_Text.D1>
-
-<Unstyled>
-    <ThemeProvider
-        theme={{ ...V2_BaseTheme, textStyleScheme: "plusJakartaSans" }}
-    >
-        <FontDisplay />
-    </ThemeProvider>
-</Unstyled>
-
-## Text Styles
-
-<Unstyled>
-    <ThemeProvider
-        theme={{ ...V2_BaseTheme, textStyleScheme: "plusJakartaSans" }}
-    >
-        <FontSizeDisplay textStyles={BasePlusSansJakartaStyleSet} />
-    </ThemeProvider>
-</Unstyled>
